Fixed sizing.
Tightened up the sizing of the graph and the table to be exact, but
this currently works by calling the calculate resize after finding a query
which increases the size of the controlbox. Its not a lot of overhead,
but i'll do a better solution later. (its ok because if theres no changes
react doesn't have to re-render)