Initial Menu Bar and DTW Bias
Added locaction and menu bar - it should be easy to add more options and
icons to it. Added a first draft for the bias. Currently moves only the result
(not the query) - which i think is probably correct. It can move the result
of the graph but its ok. Its currently not 'very' styled.