Similarity Store
Moved the functionality of the similarity store out of the main store.
Currently there is a problem with accessing the main store from within
the similarity store, but not vice versa. One of the main functional
changes I added was to split the query information (start/end/seq) into
different variables for different states. This isn't fully tested
because of the aforementioned issue. Commiting as a stash.