Connected Scatter Plot
Added initial scatter plot and enabled ability to see it via the
menu bar. Currently there is no indication of direction and its
rather mangled - but this largely a trait of the chart itself. The
expected data is similar as other charts but will look for the color
and the strokeWidth directly within the data as it does not draw more
than one line. Later we could separate this out.