Best Distinct Match
Defined a declartive FIND_DISTINCT, when set to 1 will call cause
the similarity function to find the best *distinct* interval that is
from a different ts than the query. The declaritive is defined in
GroupableTimeSeriesSet.h, and changes the function
GroupableTimeSeriesSet::similar. If no ts in the group then the function
will return the best match.