Horizon Overview Chart with View Selection
NOTE: If we were doing dev branches, this would be on dev not master,
but I'd rather not get split apart at this point. The over view chart
is now an overlapping area chart. Not quite a horizon. Currently
selection works on it pretty well. You can select what you want to view
on the top chart. Sadly, sliding does not fully work yet as that causes
a function to change the view which also updates the chart and thus
hides the brush. I shall hopefully find a fix soon. Currently things are
not yet perfect as some of our logic for the slicing of arrays in
ViewGraphs needs to be further updated - but basically, what we are
viewing is not necessarily what we picked for a query.