Update ViewGraphs and OverviewChart
Generally: made necessary changes to have the right of the chart ready
to go when we switch schemes.
Key Finding: the equality in the d3 for when to update vs when to
destory and create is a bit flawed. a = [5] and b = [5], a == b, is
false, so every single time the d3 components will have to be completely
re rendered. So far I only changed it in the overview chart.